[T2] Refactor: organize core files
