+++ lib/core/ui/ui.dart
@@ -1,0 +1,4 @@
+//GENERATED BARREL FILE 
+export 'app.dart'; 
+export 'colors.dart'; 
+export 'theme.dart';
