+++ lib/core/widgets/app_error_widget.dart
