+++ lib/feed/presentation/widgets/feed_widget.dart
@@ -2,10 +2,10 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:gripable_assignment/core/providers/subreddit_info_provider.dart';
 import 'package:gripable_assignment/core/types/sort_type.dart';
-import 'package:gripable_assignment/core/ui/app_bottom_loader.dart';
-import 'package:gripable_assignment/core/ui/app_empty_data_widget.dart';
-import 'package:gripable_assignment/core/ui/app_error_widget.dart';
-import 'package:gripable_assignment/core/ui/app_progress_widget.dart';
+import 'package:gripable_assignment/core/widgets/app_bottom_loader.dart';
+import 'package:gripable_assignment/core/widgets/app_empty_data_widget.dart';
+import 'package:gripable_assignment/core/widgets/app_error_widget.dart';
+import 'package:gripable_assignment/core/widgets/app_progress_widget.dart';
 import 'package:gripable_assignment/feed/data/data.dart';
 import 'package:gripable_assignment/feed/domain/repository/feed_repository.dart';
 import 'package:gripable_assignment/feed/presentation/blocs/feed/feed_bloc.dart';
