+++ lib/main_production.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
-import 'package:gripable_assignment/app/app.dart';
 import 'package:gripable_assignment/bootstrap/bootstrap.dart';
+import 'package:gripable_assignment/core/ui/ui.dart';
 import 'package:gripable_assignment/gen/assets.gen.dart';
 
 Future<void> main() async {
