+++ test/core/ui/app_test.dart
@@ -1,9 +1,10 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:gripable_assignment/app/app.dart';
 import 'package:gripable_assignment/core/types/sort_type.dart';
+import 'package:gripable_assignment/core/ui/ui.dart';
 import 'package:gripable_assignment/feed/data/data.dart';
 import 'package:gripable_assignment/feed/presentation/pages/feed_home_page.dart';
 import 'package:mocktail/mocktail.dart';
+
 import '../../helpers/helpers.dart';
 import '../../helpers/mocks.dart';
